refactor: split config loader into per-concern modules
`loader.ts` had grown to ~360 lines mixing markdown loaders, two
validation-heavy JSON loaders, and orchestration, with the JSON validators
only reachable through the whole-config `loadConfig` funnel. Split by concern:

- `json-config.ts` — config-path resolution plus a shared `loadJsonConfig`
  helper absorbing the read/ENOENT/parse boilerplate the two JSON loaders
  duplicated.
- `messages.ts` — the markdown message-list and instruction loaders.
- `profiles.ts` / `tool-roles.ts` — each exposes a pure validator
  (`parseProfiles` / `parseToolRoles`) plus its thin loader, so the branching
  and fail-closed permission logic are unit-tested directly with no fs mocking.
- `loader.ts` — now just `loadConfig` orchestration.

Behaviour is identical: `loadConfig`'s signature and output are unchanged, and
the existing `loader.test.ts` integration test passed untouched throughout.
Adds focused `profiles.test.ts` and `tool-roles.test.ts`.
